Rename headers to avoid clash.
See e.g: https://en.wikipedia.org/wiki/Assert.h , which clashes with our
assert.h and other too generic, non namespaced-prefixed, header
basenames.

We have placed all header files under include/OB_table/ per the discussion at:
https://github.com/Veltas/OB_table/pull/1 .
