+++ include/OB_table/hash.h
@@ -2,7 +2,7 @@
 #define VV_HASH_INCLUDED
 
 #include <stddef.h>
-#include "util.h"
+#include <OB_table/util.h>
 
 size_t FNV_1a_hash(union param, const void *);
