+++ source/hash.c
@@ -1,4 +1,4 @@
-#include "hash.h"
+#include <OB_table/hash.h>
 
 #include <limits.h>
 #include <stdint.h>
